chore: Add files from old project
Prior to this project, I was working on a fork of the SSDBot by
ocmarin. It was supposed to be a refactoring of the bot. Instead, I am
choosing to write my own bot from scratch.

The idea for doing so is I would like to write more bots in the future
for fun. I decided that writing more modular code would be more
beneficial in terms in maintenance as well as reusability. Additionally,
opening the source code to others would benefit the greater population
of developers, too.