chore: Add models. Add SsdDataService. Update main.py.
Added two model class -- reddit_post and ssd_spec. `reddit_post`
contains the data related to one specific Reddit post retrieved from
the reddit delegator and is used by the reddit processor. The same idea
applies to `ssd_spec` which is a wrapper around a map of the data
relevant to one SSD result.

A `services/` directory with SsdDataService was added. This class holds
the data from the Newmaxx spreadsheet and is represented as a Singleton.
The reasoning behind this is the separatation of concerns, where the
classes can individually retrieve external data (delegator),
store and access a local copy of that data (service), and process that
local data (processor).

As with all of the new changes, the `main.py` class needed to be
updated. Since this application isn't in v1 yet, I found this to be
acceptable.

Of the other classes that were modified per this commit, there isn't
much to say about the changes other than they are mostly method stubs
added. The intention is to write these out further and I branch out
into writing the other classes (and test cases!)