Refactor setup script and project file structure
I did not like how the project was structured, so I modified it so that
`bot/` is the primary folder in which all of the code lives.

The `setup` script needed modification to ensure the environment is
setup correctly. That includes appending .template to praw.ini so That
my own private secrets in `praw.ini` are not accidentally uploaded to the
repo again. The script also guarantees that `praw.ini` is moved to the
`$HOME/.config/` folder so that PRAW can find the appropriate site.

The Reddit client is now the primary interaction point with the Reddit
API, as it always should have been. A child class named BuildAPCSales
has been added but not built out.

Exception classes have been added for customization purposes. It was fun
until I realized I don't know how to appropriately import modules and
packages in Python. That was a fun learning curve.

I created a RedditFactory module, but honestly, I hate it and am
over-engineering for a problem I don't have at this moment. I
anticipate it will be deleted by the next commit.

README was updated. The copy-paste work from ocmarin's project is gone.
Now it better reflects my own work and intentions.